refactor(gateway): extract env/bootstrap helpers from run.py (slice 11 of #54962)
